+++ src/PVOutput.Net/Objects/Core/FormatHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Globalization;
+using System.Net;
 using System.Reflection;
 using PVOutput.Net.Enums;
 
@@ -70,6 +71,17 @@ namespace PVOutput.Net.Objects.Core
             }
 
             return (string)Convert.ChangeType(value, typeof(string), CultureInfo.CreateSpecificCulture("en-US"));
+        }
+
+        internal static string UrlEncode(string value)
+        {
+            if (value == null)
+            {
+                return null;
+            }
+
+            // + is only a valid character for encoding space in application/x-www-form-urlencoded content
+            return WebUtility.UrlEncode(value).Replace("+", "%20");
         }
 
         internal static TResultType? GetValue<TResultType>(string value) where TResultType : struct