+++ src/PVOutput.Net/Requests/Modules/AddBatchOutputRequest.cs
@@ -6,7 +6,6 @@ using PVOutput.Net.Enums;
 using PVOutput.Net.Objects.Core;
 using PVOutput.Net.Objects;
 using PVOutput.Net.Requests.Base;
-using System.Net;
 
 namespace PVOutput.Net.Requests.Modules
 {
@@ -91,7 +90,7 @@ namespace PVOutput.Net.Requests.Modules
 
             if (output.Comments != null)
             {
-                sb.Append(WebUtility.UrlEncode(output.Comments));
+                sb.Append(FormatHelper.UrlEncode(output.Comments));
             }
             sb.Append(',');