+++ src/PVOutput.Net/Requests/Modules/AddOutputRequest.cs
@@ -5,7 +5,6 @@ using PVOutput.Net.Enums;
 using PVOutput.Net.Objects.Core;
 using PVOutput.Net.Objects;
 using PVOutput.Net.Requests.Base;
-using System.Net;
 
 namespace PVOutput.Net.Requests.Modules
 {
@@ -27,7 +26,7 @@ namespace PVOutput.Net.Requests.Modules
             ["cd"] = FormatHelper.GetEnumerationDescription(Output.Condition),
             ["tm"] = FormatHelper.GetValueAsString(Output.MinimumTemperature),
             ["tx"] = FormatHelper.GetValueAsString(Output.MaximumTemperature),
-            ["cm"] = WebUtility.UrlEncode(Output.Comments),
+            ["cm"] = FormatHelper.UrlEncode(Output.Comments),
             ["ip"] = FormatHelper.GetValueAsString(Output.PeakEnergyImport),
             ["io"] = FormatHelper.GetValueAsString(Output.OffPeakEnergyImport),
             ["is"] = FormatHelper.GetValueAsString(Output.ShoulderEnergyImport),